hrvdata/requirements.txt
Django==1.8.2
django-appconf==1.0.1
django-bootstrap-form==3.2
django-jsonfield==0.9.13
django-user-accounts==1.1
hrv==0.0.1
metron==1.3.5
numpy==1.9.2
Pillow==2.9.0
pinax-eventlog==1.0.0
pinax-theme-bootstrap==5.7.6
pytz==2015.4
reportlab==3.2.0
scipy==0.16.0
six==1.9.0
wheel==0.24.0
